moved some SimpleSearch specific code out of the submit function
